refactor: extract F56 to its own file with implementation and tests
